Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.932224  3.868259  5.985291  7.156070  9.984125 10.752037 13.786166
#>  [8] 14.675521 17.120172 19.042498
#> 
#> $finite_difference
#>  [1]  0.9322243  3.8682594  5.9852918  7.1560707  9.9841248 10.7520370
#>  [7] 13.7861666 14.6755209 17.1201723 19.0424980
#> 
#> $relative_error
#>  [1] 3.426682e-07 1.400020e-07 1.204099e-07 1.239625e-07 1.388944e-08
#>  [6] 2.802100e-08 1.975015e-08 1.313642e-08 9.598425e-09 8.196911e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.253242e+00 ~ 2.153242e+00   [4.64414e-02]
#> 
#> $analytic
#>  [1]  1.447600  2.253242  5.360411  6.945854  8.357966 10.113643 12.160914
#>  [8] 15.457229 17.931419 18.932858
#> 
#> $finite_difference
#>  [1]  1.447601  2.153242  5.360411  6.945854  8.357966 10.113643 12.160914
#>  [8] 15.457229 17.931419 18.932858
#> 
#> $relative_error
#>  [1] 7.253836e-07 4.644140e-02 5.284618e-08 5.182317e-08 1.987899e-08
#>  [6] 1.992122e-08 1.949591e-08 2.259434e-09 1.410117e-08 1.670520e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.393461e+00 ~ 1.393459e+00   [9.661041e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.536430e+00 ~ 3.436430e+00   [2.910008e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.106937e+00 ~ 4.106936e+00   [1.847783e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.763662e+00 ~ 6.763661e+00   [8.103353e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 9.960505e+00 ~ 9.960505e+00   [7.723227e-09]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.161182e+01 ~ 1.161182e+01   [4.237480e-09]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.313129e+01 ~ 1.313129e+01   [1.168218e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.477770e+01 ~ 1.477770e+01   [4.474573e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.619365e+01 ~ 1.619365e+01   [9.291319e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.865286e+01 ~ 1.865286e+01   [5.197632e-09]
#> 
#> $analytic
#>          [,1]    [,2]     [,3]     [,4]     [,5]     [,6]     [,7]    [,8]
#> [1,] 1.000000 1.01000 1.000000 1.000000 1.000000  1.00000  1.00000  1.0000
#> [2,] 1.393461 3.53643 4.106937 6.763662 9.960505 11.61182 13.13129 14.7777
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.19365 18.65286
#> 
#> $finite_difference
#>          [,1]    [,2]     [,3]     [,4]     [,5]     [,6]     [,7]    [,8]
#> [1,] 1.000000 1.00000 1.000000 1.000000 1.000000  1.00000  1.00000  1.0000
#> [2,] 1.393459 3.43643 4.106936 6.763661 9.960505 11.61182 13.13129 14.7777
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.19365 18.65286
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]        [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00
#> [2,] 9.661041e-07 0.02910008 1.847783e-07 8.103353e-08 7.723227e-09 4.23748e-09
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.168218e-08 4.474573e-09 9.291319e-09 5.197632e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>