R interface to NLopt

is.nloptr preforms checks to see if a fully specified problem is supplied to nloptr. Mostly for internal use.

Usage

is.nloptr(x)

Arguments

x

object to be tested.

Value

Logical. Return TRUE if all tests were passed, otherwise return FALSE or exit with Error.

See also

Author

Jelmer Ypma