nloptr.get.default.options {nloptr}	R Documentation

Return a data.frame with all the options that can be supplied to nloptr.

Description

This function returns a data.frame with all the options that can be supplied to nloptr. The data.frame contains the default values of the options and an explanation. A user-friendly way to show these options is by using the function nloptr.print.options.

Usage

nloptr.get.default.options()

Value

The return value contains a data.frame with the following elements

name
name of the option

type
type (numeric, logical, integer, character)

possible_values
string explaining the values the option can take

default
default value of the option (as a string)

is_termination_condition
is this option part of the termination conditions?

description
description of the option (taken from NLopt website if it's an option that is passed on to NLopt).

Author(s)

Jelmer Ypma

See Also

nloptr nloptr.print.options

[Package nloptr version 2.0.3 Index]