Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.09394402  3.30080713  5.75866633  6.72037900  8.06944994 11.08274694
#>  [7] 12.35275675 14.14823005 17.32256848 18.70419458
#> 
#> $finite_difference
#>  [1]  0.09394455  3.30080605  5.75866572  6.72037888  8.06944962 11.08274714
#>  [7] 12.35275650 14.14823008 17.32256868 18.70419464
#> 
#> $relative_error
#>  [1] 5.630888e-06 3.275762e-07 1.066310e-07 1.794635e-08 3.966758e-08
#>  [6] 1.803920e-08 2.058254e-08 1.546913e-09 1.166370e-08 3.082136e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.54759e+00 ~ 3.44759e+00   [2.90058e-02]
#> 
#> $analytic
#>  [1]  1.125923  3.547590  4.383803  6.955286  8.937947 10.134981 13.876456
#>  [8] 14.324363 17.242012 18.260344
#> 
#> $finite_difference
#>  [1]  1.125923  3.447590  4.383803  6.955286  8.937947 10.134981 13.876457
#>  [8] 14.324363 17.242012 18.260345
#> 
#> $relative_error
#>  [1] 6.038294e-08 2.900580e-02 8.887335e-09 6.641510e-08 3.393752e-08
#>  [6] 5.265405e-08 1.425241e-08 2.044775e-08 1.669655e-08 1.939624e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.634826e+00 ~ 1.634827e+00   [1.882781e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.380722e+00 ~ 2.280722e+00   [4.384608e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.109800e+00 ~ 4.109800e+00   [1.394030e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.022178e+00 ~ 7.022178e+00   [2.002651e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 9.271832e+00 ~ 9.271832e+00   [4.379465e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.088479e+01 ~ 1.088479e+01   [2.534057e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.397557e+01 ~ 1.397557e+01   [1.368010e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.529743e+01 ~ 1.529743e+01   [7.975407e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.663762e+01 ~ 1.663762e+01   [5.112559e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.909688e+01 ~ 1.909688e+01   [9.938979e-09]
#> 
#> $analytic
#>          [,1]     [,2]   [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.0000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.634826 2.380722 4.1098 7.022178 9.271832 10.88479 13.97557 15.29743
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.63762 19.09688
#> 
#> $finite_difference
#>          [,1]     [,2]   [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.0000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.634827 2.280722 4.1098 7.022178 9.271832 10.88479 13.97557 15.29743
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.63762 19.09688
#> 
#> $relative_error
#>              [,1]       [,2]        [,3]         [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.00000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.882781e-07 0.04384608 1.39403e-07 2.002651e-08 4.379465e-08 2.534057e-08
#>             [,7]         [,8]         [,9]        [,10]
#> [1,] 0.00000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.36801e-08 7.975407e-09 5.112559e-09 9.938979e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>