Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.493297  2.328725  4.209793  6.535193  8.721033 10.536435 12.603592
#>  [8] 14.296950 17.357856 19.674624
#> 
#> $finite_difference
#>  [1]  1.493298  2.328726  4.209794  6.535193  8.721033 10.536435 12.603592
#>  [8] 14.296950 17.357856 19.674625
#> 
#> $relative_error
#>  [1] 2.678656e-07 5.031094e-07 2.368979e-07 1.118694e-07 6.408491e-08
#>  [6] 5.819052e-09 6.771804e-09 3.325466e-08 3.776662e-09 2.015577e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.547099e+00 ~ 2.447098e+00   [4.086506e-02]
#> 
#> $analytic
#>  [1]  0.2987444  2.5470986  4.8178349  6.5252282  9.6739778 10.6964521
#>  [7] 13.2499002 15.5738987 16.6036436 18.0219705
#> 
#> $finite_difference
#>  [1]  0.2987442  2.4470978  4.8178342  6.5252285  9.6739777 10.6964518
#>  [7] 13.2499003 15.5738988 16.6036436 18.0219707
#> 
#> $relative_error
#>  [1] 7.435138e-07 4.086506e-02 1.453993e-07 4.103385e-08 1.288105e-08
#>  [6] 2.932753e-08 3.293543e-09 5.740821e-09 8.725347e-11 1.206659e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 9.301413e-01 ~ 9.301414e-01   [1.161473e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.727067e+00 ~ 3.627068e+00   [2.757045e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.198555e+00 ~ 4.198554e+00   [1.953863e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.598886e+00 ~ 7.598886e+00   [6.936920e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.583171e+00 ~ 8.583170e+00   [5.635784e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.113003e+01 ~ 1.113003e+01   [5.634224e-09]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.370891e+01 ~ 1.370891e+01   [2.668898e-10]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.438302e+01 ~ 1.438302e+01   [3.917468e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.600359e+01 ~ 1.600359e+01   [5.522020e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.939361e+01 ~ 1.939361e+01   [5.388084e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.9301413 3.727067 4.198555 7.598886 8.583171 11.13003 13.70891 14.38302
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.00359 19.39361
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]    [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.00000  1.00000  1.00000  1.00000
#> [2,] 0.9301414 3.627068 4.198554 7.598886 8.58317 11.13003 13.70891 14.38302
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.00359 19.39361
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]        [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.00000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.161473e-07 0.02757045 1.953863e-07 6.93692e-08 5.635784e-08 5.634224e-09
#>              [,7]         [,8]        [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.00000e+00 0.000000e+00
#> [2,] 2.668898e-10 3.917468e-09 5.52202e-09 5.388084e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>