Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.473863  2.029237  4.180212  6.089490  8.678308 11.682239 12.320662
#>  [8] 15.307603 16.588518 19.753948
#> 
#> $finite_difference
#>  [1]  1.473862  2.029236  4.180211  6.089490  8.678308 11.682239 12.320662
#>  [8] 15.307603 16.588518 19.753948
#> 
#> $relative_error
#>  [1] 8.053474e-07 4.557397e-07 1.001827e-07 7.035210e-09 2.292273e-08
#>  [6] 1.300785e-08 1.551221e-08 1.794795e-08 1.019300e-08 1.303120e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.258566e+00 ~ 3.158566e+00   [3.166e-02]
#> 
#> $analytic
#>  [1]  0.695663  3.258566  4.369749  7.108420  8.166591 10.218133 13.307885
#>  [8] 14.340787 17.060495 18.216128
#> 
#> $finite_difference
#>  [1]  0.6956635  3.1585655  4.3697497  7.1084204  8.1665909 10.2181327
#>  [7] 13.3078859 14.3407869 17.0604952 18.2161285
#> 
#> $relative_error
#>  [1] 6.700466e-07 3.166000e-02 5.065378e-08 7.513814e-08 2.533981e-08
#>  [6] 1.429443e-08 3.183716e-08 1.376775e-08 1.309843e-08 1.535322e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.023964e+00 ~ 1.023964e+00   [1.364290e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.408311e+00 ~ 2.308311e+00   [4.332165e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.665659e+00 ~ 5.665658e+00   [4.112247e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.840941e+00 ~ 6.840941e+00   [5.105233e-09]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.600331e+00 ~ 8.600331e+00   [5.067932e-09]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.090021e+01 ~ 1.090021e+01   [3.616994e-09]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.289715e+01 ~ 1.289715e+01   [1.739265e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.492358e+01 ~ 1.492358e+01   [3.057900e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.788653e+01 ~ 1.788653e+01   [1.168646e-08]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.822915e+01 ~ 1.822915e+01   [5.267352e-09]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.023964 2.408311 5.665659 6.840941 8.600331 10.90021 12.89715 14.92358
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.88653 18.22915
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.023964 2.308311 5.665658 6.840941 8.600331 10.90021 12.89715 14.92358
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.88653 18.22915
#> 
#> $relative_error
#>             [,1]       [,2]         [,3]         [,4]         [,5]         [,6]
#> [1,] 0.00000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.36429e-07 0.04332165 4.112247e-08 5.105233e-09 5.067932e-09 3.616994e-09
#>              [,7]       [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.0000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.739265e-08 3.0579e-09 1.168646e-08 5.267352e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>