Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

Usage

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.5913913  3.8054079  4.4223155  6.8425853  8.1617139 10.9530364
#>  [7] 13.2636155 15.2651858 16.6956614 19.3240802
#> 
#> $finite_difference
#>  [1]  0.5913887  3.8054066  4.4223150  6.8425846  8.1617134 10.9530366
#>  [7] 13.2636152 15.2651858 16.6956613 19.3240803
#> 
#> $relative_error
#>  [1] 4.388197e-06 3.563366e-07 1.296570e-07 9.459425e-08 6.416333e-08
#>  [6] 1.660894e-08 1.975587e-08 5.307833e-09 6.892215e-09 3.889330e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 3.471044e+00 ~ 3.371044e+00   [2.966441e-02]
#> 
#> $analytic
#>  [1]  1.860670  3.471044  4.602297  7.726301  9.837235 10.062662 12.189008
#>  [8] 14.983514 17.767496 18.867624
#> 
#> $finite_difference
#>  [1]  1.860668  3.371044  4.602296  7.726300  9.837234 10.062662 12.189008
#>  [8] 14.983514 17.767496 18.867624
#> 
#> $relative_error
#>  [1] 8.586614e-07 2.966441e-02 1.311633e-07 3.501666e-08 5.140755e-08
#>  [6] 3.447573e-08 1.888338e-09 2.641651e-09 5.008756e-10 7.818764e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 6.684885e-01 ~ 6.684875e-01   [1.366708e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 4.042231e+00 ~ 3.942232e+00   [2.536607e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 4.831071e+00 ~ 4.831071e+00   [2.936646e-08]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 6.148032e+00 ~ 6.148032e+00   [9.937287e-08]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 8.989516e+00 ~ 8.989516e+00   [3.266538e-08]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.136968e+01 ~ 1.136968e+01   [1.799351e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.231104e+01 ~ 1.231104e+01   [2.369986e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.561847e+01 ~ 1.561847e+01   [1.964794e-08]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.672423e+01 ~ 1.672423e+01   [1.251411e-08]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.988105e+01 ~ 1.988105e+01   [6.525473e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.6684885 4.042231 4.831071 6.148032 8.989516 11.36968 12.31104 15.61847
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.72423 19.88105
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.6684875 3.942232 4.831071 6.148032 8.989516 11.36968 12.31104 15.61847
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.72423 19.88105
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.366708e-06 0.02536607 2.936646e-08 9.937287e-08 3.266538e-08
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.799351e-08 2.369986e-08 1.964794e-08 1.251411e-08 6.525473e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>