Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.415524  3.513012  4.241184  6.401803  9.453251 10.620032 13.472249
#>  [8] 14.834758 17.068854 18.589981
#> 
#> $finite_difference
#>  [1]  1.415524  3.513014  4.241184  6.401804  9.453251 10.620032 13.472250
#>  [8] 14.834758 17.068854 18.589981
#> 
#> $relative_error
#>  [1] 7.138596e-08 4.123731e-07 3.535403e-08 8.670498e-08 3.694454e-09
#>  [6] 2.934856e-08 1.817598e-08 2.194153e-08 2.647806e-08 1.285517e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.264737e+00 ~ 3.164738e+00   [3.159802e-02]
#> 
#> $analytic
#>  [1]  0.6388522  3.2647371  4.9881006  6.1464013  9.8919144 10.9323937
#>  [7] 13.8862809 15.5426895 16.3966379 18.9604425
#> 
#> $finite_difference
#>  [1]  0.6388512  3.1647377  4.9881007  6.1464014  9.8919144 10.9323940
#>  [7] 13.8862811 15.5426898 16.3966382 18.9604427
#> 
#> $relative_error
#>  [1] 1.621812e-06 3.159802e-02 1.244726e-08 1.712212e-08 1.553473e-09
#>  [6] 3.019959e-08 1.863048e-08 2.154135e-08 2.052359e-08 1.479963e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.707084e+00 ~ 1.707085e+00   [4.192653e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.044951e+00 ~ 2.944950e+00   [3.395666e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.034598e+00 ~ 5.034598e+00   [3.952502e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.220814e+00 ~ 6.220814e+00   [2.230690e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 9.411032e+00 ~ 9.411032e+00   [1.252843e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.153068e+01 ~ 1.153068e+01   [2.557688e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.218204e+01 ~ 1.218204e+01   [1.256518e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.478248e+01 ~ 1.478248e+01   [6.772693e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.694340e+01 ~ 1.694340e+01   [3.542298e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.883155e+01 ~ 1.883155e+01   [6.216548e-09]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.707084 3.044951 5.034598 6.220814 9.411032 11.53068 12.18204 14.78248
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 16.9434 18.83155
#> 
#> $finite_difference
#>          [,1]    [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.00000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.707085 2.94495 5.034598 6.220814 9.411032 11.53068 12.18204 14.78248
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 16.9434 18.83155
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]        [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.00000e+00 0.000000e+00 0.000000e+00
#> [2,] 4.192653e-07 0.03395666 3.952502e-08 2.23069e-08 1.252843e-08 2.557688e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.256518e-08 6.772693e-09 3.542298e-09 6.216548e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>