Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.3613055  3.4193679  4.3988807  6.7087348  8.2700853 11.0339364
#>  [7] 13.0856968 14.5142574 17.5593287 18.3315350
#> 
#> $finite_difference
#>  [1]  0.3613052  3.4193687  4.3988813  6.7087355  8.2700851 11.0339368
#>  [7] 13.0856972 14.5142574 17.5593287 18.3315353
#> 
#> $relative_error
#>  [1] 7.900532e-07 2.411808e-07 1.201147e-07 1.041862e-07 1.737625e-08
#>  [6] 3.395906e-08 3.216931e-08 3.464967e-09 3.429831e-09 2.001693e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.426403e+00 ~ 3.326403e+00   [3.006252e-02]
#> 
#> $analytic
#>  [1]  0.5376526  3.4264027  4.7219576  7.0254019  8.5625123 11.3561770
#>  [7] 13.1105299 14.7570912 17.1503965 19.1686042
#> 
#> $finite_difference
#>  [1]  0.5376511  3.3264027  4.7219569  7.0254021  8.5625122 11.3561770
#>  [7] 13.1105298 14.7570910 17.1503966 19.1686043
#> 
#> $relative_error
#>  [1] 2.859873e-06 3.006252e-02 1.597253e-07 1.855910e-08 6.177991e-09
#>  [6] 1.571862e-09 8.438089e-09 9.687412e-09 3.369820e-09 1.656777e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.286475e+00 ~ 1.286472e+00   [2.336863e-06]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.873962e+00 ~ 2.773962e+00   [3.604940e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.610510e+00 ~ 5.610509e+00   [9.132551e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.303620e+00 ~ 7.303619e+00   [3.334523e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.116095e+00 ~ 8.116095e+00   [1.969111e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.092820e+01 ~ 1.092820e+01   [3.775333e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.254876e+01 ~ 1.254876e+01   [1.281291e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.436100e+01 ~ 1.436100e+01   [5.804147e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.715571e+01 ~ 1.715571e+01   [6.194687e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.913352e+01 ~ 1.913352e+01   [6.167113e-09]
#> 
#> $analytic
#>          [,1]     [,2]    [,3]    [,4]     [,5]    [,6]     [,7]   [,8]
#> [1,] 1.000000 1.010000 1.00000 1.00000 1.000000  1.0000  1.00000  1.000
#> [2,] 1.286475 2.873962 5.61051 7.30362 8.116095 10.9282 12.54876 14.361
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.15571 19.13352
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]    [,6]     [,7]   [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.0000  1.00000  1.000
#> [2,] 1.286472 2.773962 5.610509 7.303619 8.116095 10.9282 12.54876 14.361
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.15571 19.13352
#> 
#> $relative_error
#>              [,1]      [,2]         [,3]         [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.0100000 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 2.336863e-06 0.0360494 9.132551e-08 3.334523e-08 1.969111e-08 3.775333e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.281291e-08 5.804147e-09 6.194687e-09 6.167113e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>