Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

Usage

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.534992  3.889084  4.910609  7.332501  8.733903 11.847338 13.237700
#>  [8] 15.524436 17.197161 18.690499
#> 
#> $finite_difference
#>  [1]  1.534992  3.889084  4.910609  7.332500  8.733904 11.847338 13.237700
#>  [8] 15.524436 17.197161 18.690499
#> 
#> $relative_error
#>  [1] 2.011303e-07 3.244831e-08 1.106322e-07 6.014064e-08 2.670091e-08
#>  [6] 9.420132e-09 1.170888e-08 6.359020e-09 5.743502e-09 1.060354e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 3.501141e+00 ~ 3.401142e+00   [2.940164e-02]
#> 
#> $analytic
#>  [1]  1.091364  3.501141  4.778033  6.832459  9.236990 10.672067 12.144026
#>  [8] 15.548590 17.747276 19.432702
#> 
#> $finite_difference
#>  [1]  1.091366  3.401142  4.778033  6.832458  9.236990 10.672067 12.144026
#>  [8] 15.548591 17.747276 19.432702
#> 
#> $relative_error
#>  [1] 1.340621e-06 2.940164e-02 1.601573e-08 1.335824e-08 6.916618e-09
#>  [6] 8.959444e-09 1.411640e-08 1.802916e-08 1.647193e-08 1.179927e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 3.991215e-01 ~ 3.991203e-01   [2.933132e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 3.079515e+00 ~ 2.979515e+00   [3.356260e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 4.634691e+00 ~ 4.634691e+00   [1.071377e-07]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 6.372316e+00 ~ 6.372316e+00   [3.069178e-08]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 8.467413e+00 ~ 8.467413e+00   [1.777423e-08]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.047034e+01 ~ 1.047034e+01   [1.728569e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.315985e+01 ~ 1.315985e+01   [7.537004e-09]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.426017e+01 ~ 1.426017e+01   [1.113525e-08]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.777830e+01 ~ 1.777830e+01   [9.603983e-10]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.848172e+01 ~ 1.848172e+01   [1.508723e-08]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3991215 3.079515 4.634691 6.372316 8.467413 10.47034 13.15985 14.26017
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.7783 18.48172
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3991203 2.979515 4.634691 6.372316 8.467413 10.47034 13.15985 14.26017
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.7783 18.48172
#> 
#> $relative_error
#>              [,1]      [,2]         [,3]         [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.0100000 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 2.933132e-06 0.0335626 1.071377e-07 3.069178e-08 1.777423e-08 1.728569e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 7.537004e-09 1.113525e-08 9.603983e-10 1.508723e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>