Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.5961005  2.6332004  5.5271023  6.9497546  9.2178063 10.6873649
#>  [7] 12.3584375 14.5996995 17.5115225 19.5174870
#> 
#> $finite_difference
#>  [1]  0.5960999  2.6331997  5.5271022  6.9497547  9.2178062 10.6873652
#>  [7] 12.3584377 14.5996995 17.5115225 19.5174873
#> 
#> $relative_error
#>  [1] 1.135055e-06 2.871920e-07 3.055486e-08 1.005060e-08 5.647866e-09
#>  [6] 2.942508e-08 1.373691e-08 3.093841e-09 7.741142e-10 1.642431e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 2.191398e+00 ~ 2.0914e+00   [4.781406e-02]
#> 
#> $analytic
#>  [1]  0.1011055  2.1913985  4.1190495  7.6971019  8.5951266 10.4983047
#>  [7] 13.3973163 15.8158946 17.3880026 19.7528536
#> 
#> $finite_difference
#>  [1]  0.1011086  2.0914001  4.1190503  7.6971025  8.5951271 10.4983050
#>  [7] 13.3973165 15.8158951 17.3880030 19.7528538
#> 
#> $relative_error
#>  [1] 2.984862e-05 4.781406e-02 2.070711e-07 8.947692e-08 6.344169e-08
#>  [6] 2.729364e-08 1.516433e-08 3.250465e-08 2.330806e-08 1.094323e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 1.476262e+00 ~ 1.476265e+00   [1.791339e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 2.139456e+00 ~ 2.039457e+00   [4.903185e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 4.738552e+00 ~ 4.738553e+00   [2.043704e-07]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 7.601455e+00 ~ 7.601455e+00   [3.369272e-09]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 8.486872e+00 ~ 8.486872e+00   [5.833611e-08]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.031853e+01 ~ 1.031853e+01   [5.180769e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.282291e+01 ~ 1.282291e+01   [1.972412e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.549696e+01 ~ 1.549696e+01   [3.293323e-08]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.663655e+01 ~ 1.663655e+01   [2.332211e-08]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.975928e+01 ~ 1.975928e+01   [1.854229e-08]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.476262 2.139456 4.738552 7.601455 8.486872 10.31853 12.82291 15.49696
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.63655 19.75928
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.476265 2.039457 4.738553 7.601455 8.486872 10.31853 12.82291 15.49696
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.63655 19.75928
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.791339e-06 0.04903185 2.043704e-07 3.369272e-09 5.833611e-08
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 5.180769e-08 1.972412e-08 3.293323e-08 2.332211e-08 1.854229e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>