Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.887633  3.550308  5.362713  7.522374  9.892164 10.287010 12.768978
#>  [8] 14.359173 17.141781 18.983213
#> 
#> $finite_difference
#>  [1]  1.887634  3.550308  5.362713  7.522373  9.892165 10.287010 12.768978
#>  [8] 14.359173 17.141781 18.983213
#> 
#> $relative_error
#>  [1] 5.807092e-07 6.938407e-08 4.966857e-08 4.296103e-08 2.561753e-08
#>  [6] 1.726179e-08 1.586365e-08 7.912806e-09 2.547500e-09 1.287833e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 4.039115e+00 ~ 3.939116e+00   [2.53864e-02]
#> 
#> $analytic
#>  [1]  1.129471  4.039115  5.778031  7.858484  8.905222 10.348568 12.895201
#>  [8] 14.544919 17.274172 19.054530
#> 
#> $finite_difference
#>  [1]  1.129471  3.939116  5.778030  7.858484  8.905222 10.348569 12.895202
#>  [8] 14.544919 17.274172 19.054530
#> 
#> $relative_error
#>  [1] 3.421946e-07 2.538640e-02 4.625964e-08 4.770097e-08 1.610557e-09
#>  [6] 3.353825e-08 6.768272e-09 1.094926e-08 6.035391e-09 1.233159e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 3.376089e-01 ~ 3.376083e-01   [1.711704e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 2.611416e+00 ~ 2.511415e+00   [3.981829e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 5.986319e+00 ~ 5.986318e+00   [9.741600e-08]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 7.369892e+00 ~ 7.369892e+00   [2.280952e-08]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 8.367903e+00 ~ 8.367903e+00   [8.558740e-09]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.054072e+01 ~ 1.054072e+01   [7.966649e-09]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.217711e+01 ~ 1.217711e+01   [2.267128e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.473419e+01 ~ 1.473419e+01   [5.593932e-09]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.715225e+01 ~ 1.715225e+01   [1.023808e-08]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.870543e+01 ~ 1.870543e+01   [2.469528e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3376089 2.611416 5.986319 7.369892 8.367903 10.54072 12.17711 14.73419
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.15225 18.70543
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3376083 2.511415 5.986318 7.369892 8.367903 10.54072 12.17711 14.73419
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.15225 18.70543
#> 
#> $relative_error
#>              [,1]       [,2]       [,3]         [,4]        [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.0000e+00 0.000000e+00 0.00000e+00 0.000000e+00
#> [2,] 1.711704e-06 0.03981829 9.7416e-08 2.280952e-08 8.55874e-09 7.966649e-09
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 2.267128e-08 5.593932e-09 1.023808e-08 2.469528e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>