Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

Usage

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.264758  3.167043  5.579219  6.484902  9.860605 11.783110 12.811994
#>  [8] 14.159786 17.079334 19.388130
#> 
#> $finite_difference
#>  [1]  1.264759  3.167044  5.579219  6.484902  9.860606 11.783110 12.811995
#>  [8] 14.159787 17.079334 19.388130
#> 
#> $relative_error
#>  [1] 6.387954e-07 3.172981e-07 1.613628e-09 4.222251e-08 3.868624e-08
#>  [6] 7.996089e-09 3.587843e-08 3.716138e-08 1.340813e-08 8.069999e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 2.920606e+00 ~ 2.820606e+00   [3.545323e-02]
#> 
#> $analytic
#>  [1]  1.563446  2.920606  5.913905  7.440845  8.316543 11.658224 12.586676
#>  [8] 14.584997 17.763938 18.359896
#> 
#> $finite_difference
#>  [1]  1.563446  2.820606  5.913905  7.440845  8.316543 11.658225 12.586676
#>  [8] 14.584997 17.763938 18.359896
#> 
#> $relative_error
#>  [1] 2.087879e-07 3.545323e-02 7.446186e-08 2.259202e-08 2.841043e-08
#>  [6] 2.437838e-08 8.662435e-09 7.726435e-09 2.158274e-09 1.081477e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 3.419084e-01 ~ 3.419075e-01   [2.578173e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 2.202936e+00 ~ 2.102936e+00   [4.755288e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 5.143633e+00 ~ 5.143632e+00   [5.398701e-08]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 7.462253e+00 ~ 7.462253e+00   [6.365029e-09]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 8.066007e+00 ~ 8.066006e+00   [1.660353e-08]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.070467e+01 ~ 1.070467e+01   [1.289072e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.385511e+01 ~ 1.385511e+01   [1.503298e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.541915e+01 ~ 1.541915e+01   [8.003052e-09]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.748677e+01 ~ 1.748677e+01   [1.053339e-09]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.812786e+01 ~ 1.812786e+01   [1.629108e-08]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3419084 2.202936 5.143633 7.462253 8.066007 10.70467 13.85511 15.41915
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.48677 18.12786
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3419075 2.102936 5.143632 7.462253 8.066006 10.70467 13.85511 15.41915
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.48677 18.12786
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 2.578173e-06 0.04755288 5.398701e-08 6.365029e-09 1.660353e-08
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.289072e-08 1.503298e-08 8.003052e-09 1.053339e-09 1.629108e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>