Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

Usage

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.685583  3.985201  5.067213  7.004445  9.420466 10.534236 12.454957
#>  [8] 14.250799 17.650119 19.931517
#> 
#> $finite_difference
#>  [1]  1.685585  3.985201  5.067214  7.004445  9.420466 10.534236 12.454957
#>  [8] 14.250799 17.650119 19.931517
#> 
#> $relative_error
#>  [1] 1.130183e-06 5.912490e-08 1.359462e-07 2.094146e-08 3.609438e-08
#>  [6] 2.336948e-08 1.340613e-08 1.666484e-09 1.062058e-08 3.794161e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 3.295344e+00 ~ 3.195343e+00   [3.12957e-02]
#> 
#> $analytic
#>  [1]  1.359229  3.295344  5.608660  7.192924  9.872677 11.222597 12.048904
#>  [8] 15.420215 16.643239 18.529361
#> 
#> $finite_difference
#>  [1]  1.359230  3.195343  5.608660  7.192924  9.872677 11.222597 12.048904
#>  [8] 15.420216 16.643239 18.529361
#> 
#> $relative_error
#>  [1] 1.105887e-06 3.129570e-02 8.828368e-09 3.049198e-08 2.391350e-08
#>  [6] 2.240632e-08 7.420340e-09 1.280400e-08 1.165792e-08 8.609879e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 1.608087e+00 ~ 1.608086e+00   [5.606171e-07]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 2.138921e+00 ~ 2.038921e+00   [4.904519e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 4.516957e+00 ~ 4.516958e+00   [1.459436e-07]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 7.897107e+00 ~ 7.897108e+00   [7.984003e-08]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 8.939575e+00 ~ 8.939575e+00   [1.369964e-08]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.060835e+01 ~ 1.060835e+01   [3.950615e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.262289e+01 ~ 1.262289e+01   [2.371515e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.593754e+01 ~ 1.593754e+01   [1.431677e-08]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.754887e+01 ~ 1.754888e+01   [1.378941e-08]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.939834e+01 ~ 1.939834e+01   [1.606429e-08]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.608087 2.138921 4.516957 7.897107 8.939575 10.60835 12.62289 15.93754
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.54887 19.39834
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.608086 2.038921 4.516958 7.897108 8.939575 10.60835 12.62289 15.93754
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.54888 19.39834
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 5.606171e-07 0.04904519 1.459436e-07 7.984003e-08 1.369964e-08
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 3.950615e-08 2.371515e-08 1.431677e-08 1.378941e-08 1.606429e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>