Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.5059955  2.7711952  5.0888149  6.3172927  8.1824504 10.1864751
#>  [7] 12.4781445 15.0795308 16.8501383 18.3718422
#> 
#> $finite_difference
#>  [1]  0.5059967  2.7711964  5.0888151  6.3172932  8.1824509 10.1864751
#>  [7] 12.4781445 15.0795307 16.8501383 18.3718422
#> 
#> $relative_error
#>  [1] 2.281387e-06 4.113526e-07 3.099031e-08 7.408071e-08 6.224789e-08
#>  [6] 6.095152e-11 2.068492e-09 2.346907e-09 2.330590e-09 1.252115e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.910632e+00 ~ 3.810631e+00   [2.624257e-02]
#> 
#> $analytic
#>  [1]  1.437945  3.910632  4.691970  6.355855  9.279360 11.802785 13.957009
#>  [8] 15.350554 16.664487 19.130764
#> 
#> $finite_difference
#>  [1]  1.437946  3.810631  4.691970  6.355855  9.279360 11.802785 13.957009
#>  [8] 15.350554 16.664487 19.130764
#> 
#> $relative_error
#>  [1] 6.473516e-07 2.624257e-02 1.135048e-07 5.077260e-08 3.448540e-08
#>  [6] 1.512385e-09 1.278793e-08 2.141661e-08 1.130153e-08 1.650318e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 7.158493e-01 ~ 7.158470e-01   [3.173158e-06]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.615699e+00 ~ 2.515697e+00   [3.975083e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.479507e+00 ~ 5.479506e+00   [8.138476e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.549599e+00 ~ 7.549599e+00   [8.203476e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.585113e+00 ~ 8.585114e+00   [2.381160e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.087191e+01 ~ 1.087191e+01   [9.522893e-09]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.331588e+01 ~ 1.331588e+01   [1.359848e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.590758e+01 ~ 1.590758e+01   [1.642211e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.736205e+01 ~ 1.736205e+01   [2.324456e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.881998e+01 ~ 1.881998e+01   [4.597234e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.7158493 2.615699 5.479507 7.549599 8.585113 10.87191 13.31588 15.90758
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.36205 18.81998
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.715847 2.515697 5.479506 7.549599 8.585114 10.87191 13.31588 15.90758
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.36205 18.81998
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]        [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.00000e+00 0.000000e+00
#> [2,] 3.173158e-06 0.03975083 8.138476e-08 8.203476e-08 2.38116e-08 9.522893e-09
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.359848e-08 1.642211e-08 2.324456e-09 4.597234e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>