Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.4774001  2.2342768  5.5593939  6.5909210  9.5161829 11.2143552
#>  [7] 12.0517659 15.7764175 16.4841088 19.7770699
#> 
#> $finite_difference
#>  [1]  0.4774017  2.2342777  5.5593948  6.5909214  9.5161835 11.2143555
#>  [7] 12.0517660 15.7764177 16.4841092 19.7770702
#> 
#> $relative_error
#>  [1] 3.329066e-06 4.312146e-07 1.661821e-07 6.422260e-08 6.017856e-08
#>  [6] 2.138469e-08 4.840843e-09 1.210168e-08 2.931315e-08 1.915663e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.351963e+00 ~ 3.251963e+00   [3.075089e-02]
#> 
#> $analytic
#>  [1]  0.1955988  3.3519634  5.9863663  7.0543911  8.5722433 10.6563744
#>  [7] 12.0955059 15.7158172 17.4159541 19.8262514
#> 
#> $finite_difference
#>  [1]  0.1955986  3.2519627  5.9863663  7.0543909  8.5722435 10.6563746
#>  [7] 12.0955059 15.7158170 17.4159542 19.8262516
#> 
#> $relative_error
#>  [1] 1.087979e-06 3.075089e-02 7.778697e-10 2.765825e-08 2.272832e-08
#>  [6] 1.644498e-08 2.859903e-10 1.211871e-08 3.469950e-09 1.136306e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 8.779433e-01 ~ 8.779449e-01   [1.924288e-06]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.051624e+00 ~ 2.951624e+00   [3.387967e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.235591e+00 ~ 5.235592e+00   [1.218498e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.804485e+00 ~ 7.804485e+00   [1.073985e-09]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.676152e+00 ~ 8.676152e+00   [4.129478e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.163504e+01 ~ 1.163504e+01   [2.889610e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.210259e+01 ~ 1.210259e+01   [9.586062e-09]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.464470e+01 ~ 1.464470e+01   [1.465854e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.750584e+01 ~ 1.750584e+01   [1.330017e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.894061e+01 ~ 1.894061e+01   [5.212092e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]    [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.0000
#> [2,] 0.8779433 3.051624 5.235591 7.804485 8.676152 11.63504 12.10259 14.6447
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.50584 18.94061
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]    [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.0000
#> [2,] 0.8779449 2.951624 5.235592 7.804485 8.676152 11.63504 12.10259 14.6447
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.50584 18.94061
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]        [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00
#> [2,] 1.924288e-06 0.03387967 1.218498e-07 1.073985e-09 4.129478e-08 2.88961e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 9.586062e-09 1.465854e-08 1.330017e-09 5.212092e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>