Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.236853  2.473860  5.761412  7.446927  8.471707 11.549357 12.279803
#>  [8] 14.151069 16.202381 19.982607
#> 
#> $finite_difference
#>  [1]  1.236851  2.473860  5.761411  7.446926  8.471706 11.549356 12.279803
#>  [8] 14.151069 16.202381 19.982607
#> 
#> $relative_error
#>  [1] 1.653165e-06 4.592878e-08 1.380746e-07 6.503190e-08 1.434629e-08
#>  [6] 3.382784e-08 8.125908e-10 9.016364e-09 8.363427e-09 6.524932e-10
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 4.023498e+00 ~ 3.923498e+00   [2.548753e-02]
#> 
#> $analytic
#>  [1]  0.9155419  4.0234984  5.0819135  7.2681723  8.9292167 10.5216886
#>  [7] 12.9973770 14.8075235 17.0434935 19.4780884
#> 
#> $finite_difference
#>  [1]  0.9155426  3.9234982  5.0819130  7.2681723  8.9292168 10.5216885
#>  [7] 12.9973771 14.8075233 17.0434935 19.4780884
#> 
#> $relative_error
#>  [1] 7.939524e-07 2.548753e-02 9.694571e-08 8.328912e-09 7.467922e-09
#>  [6] 1.296738e-08 1.052812e-08 1.754777e-08 2.155396e-10 3.107901e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 3.916913e-01 ~ 3.916931e-01   [4.607952e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 4.075623e+00 ~ 3.975622e+00   [2.515341e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 4.952256e+00 ~ 4.952257e+00   [8.149277e-08]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 6.969719e+00 ~ 6.969719e+00   [4.483089e-08]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 9.275893e+00 ~ 9.275893e+00   [2.769100e-08]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.199865e+01 ~ 1.199865e+01   [1.875791e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.380023e+01 ~ 1.380023e+01   [1.454324e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.467876e+01 ~ 1.467876e+01   [1.427557e-08]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.693184e+01 ~ 1.693184e+01   [6.401878e-09]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.946722e+01 ~ 1.946722e+01   [7.310037e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3916913 4.075623 4.952256 6.969719 9.275893 11.99865 13.80023 14.67876
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.93184 19.46722
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3916931 3.975622 4.952257 6.969719 9.275893 11.99865 13.80023 14.67876
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.93184 19.46722
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]       [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.0000e+00 0.000000e+00
#> [2,] 4.607952e-06 0.02515341 8.149277e-08 4.483089e-08 2.7691e-08 1.875791e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.454324e-08 1.427557e-08 6.401878e-09 7.310037e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>