Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.3615824  2.3502888  5.7955734  6.2204171  8.4264042 10.2526412
#>  [7] 12.3442462 14.1119355 16.3821779 18.2222281
#> 
#> $finite_difference
#>  [1]  0.3615837  2.3502903  5.7955742  6.2204180  8.4264046 10.2526417
#>  [7] 12.3442465 14.1119356 16.3821784 18.2222282
#> 
#> $relative_error
#>  [1] 3.523525e-06 6.316361e-07 1.349842e-07 1.415605e-07 4.502760e-08
#>  [6] 4.278438e-08 1.932492e-08 8.513404e-09 3.182322e-08 9.393862e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.376088e+00 ~ 3.276089e+00   [3.052413e-02]
#> 
#> $analytic
#>  [1]  1.334183  3.376088  5.645310  6.723830  9.473005 11.690381 12.341761
#>  [8] 14.141326 17.224847 18.075715
#> 
#> $finite_difference
#>  [1]  1.334183  3.276089  5.645309  6.723829  9.473005 11.690381 12.341761
#>  [8] 14.141326 17.224847 18.075715
#> 
#> $relative_error
#>  [1] 3.727572e-07 3.052413e-02 1.596112e-07 6.800877e-08 3.171588e-08
#>  [6] 1.532237e-08 1.268285e-08 1.376437e-08 1.555973e-09 4.997772e-10
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.438882e+00 ~ 1.438881e+00   [6.330152e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.784695e+00 ~ 3.684694e+00   [2.713952e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.730326e+00 ~ 4.730325e+00   [1.149143e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.652212e+00 ~ 6.652211e+00   [5.488077e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.787949e+00 ~ 8.787949e+00   [8.721926e-09]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.196440e+01 ~ 1.196439e+01   [1.036585e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.337875e+01 ~ 1.337875e+01   [1.092412e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.507562e+01 ~ 1.507562e+01   [6.517444e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.782450e+01 ~ 1.782450e+01   [7.158190e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.828986e+01 ~ 1.828986e+01   [1.751654e-08]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]    [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.0000  1.00000  1.00000
#> [2,] 1.438882 3.784695 4.730326 6.652212 8.787949 11.9644 13.37875 15.07562
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.8245 18.28986
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.438881 3.684694 4.730325 6.652211 8.787949 11.96439 13.37875 15.07562
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.8245 18.28986
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 6.330152e-07 0.02713952 1.149143e-07 5.488077e-08 8.721926e-09
#>              [,6]         [,7]         [,8]        [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00 0.000000e+00
#> [2,] 1.036585e-08 1.092412e-08 6.517444e-09 7.15819e-09 1.751654e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>