Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.08828933  3.77289402  5.58896605  7.84504101  9.08540589 10.33253743
#>  [7] 13.35790385 15.04108270 17.38614969 18.79664830
#> 
#> $finite_difference
#>  [1]  0.08829117  3.77289581  5.58896637  7.84504128  9.08540649 10.33253797
#>  [7] 13.35790416 15.04108286 17.38614994 18.79664879
#> 
#> $relative_error
#>  [1] 2.085401e-05 4.745606e-07 5.765603e-08 3.395244e-08 6.609685e-08
#>  [6] 5.257871e-08 2.349090e-08 1.030944e-08 1.434998e-08 2.572494e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.599792e+00 ~ 2.499792e+00   [4.000313e-02]
#> 
#> $analytic
#>  [1]  1.558263  2.599792  5.142110  7.632475  8.439776 11.597478 13.807009
#>  [8] 15.231629 16.732606 19.769757
#> 
#> $finite_difference
#>  [1]  1.558262  2.499792  5.142109  7.632475  8.439776 11.597478 13.807009
#>  [8] 15.231629 16.732607 19.769757
#> 
#> $relative_error
#>  [1] 8.343439e-07 4.000313e-02 1.139225e-07 1.927933e-08 2.702452e-08
#>  [6] 1.430748e-08 1.511426e-08 9.721895e-09 1.460743e-08 9.614834e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 5.256821e-01 ~ 5.256844e-01   [4.361774e-06]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.452938e+00 ~ 3.352938e+00   [2.982456e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.678506e+00 ~ 4.678506e+00   [1.386813e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.551169e+00 ~ 6.551169e+00   [4.037380e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.574397e+00 ~ 8.574397e+00   [3.348652e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.132954e+01 ~ 1.132954e+01   [2.577066e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.369263e+01 ~ 1.369263e+01   [2.320815e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.451384e+01 ~ 1.451384e+01   [2.290794e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.733996e+01 ~ 1.733996e+01   [1.552208e-08]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.900619e+01 ~ 1.900619e+01   [9.878602e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.5256821 3.452938 4.678506 6.551169 8.574397 11.32954 13.69263 14.51384
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.33996 19.00619
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.5256844 3.352938 4.678506 6.551169 8.574397 11.32954 13.69263 14.51384
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.33996 19.00619
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]        [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.00000e+00 0.000000e+00 0.000000e+00
#> [2,] 4.361774e-06 0.02982456 1.386813e-08 4.03738e-08 3.348652e-08 2.577066e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 2.320815e-08 2.290794e-08 1.552208e-08 9.878602e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>