Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.522408  3.230016  4.494504  7.650364  9.889932 10.684124 12.530434
#>  [8] 15.484803 17.411199 19.907203
#> 
#> $finite_difference
#>  [1]  1.522408  3.230017  4.494504  7.650364  9.889932 10.684125 12.530435
#>  [8] 15.484803 17.411200 19.907204
#> 
#> $relative_error
#>  [1] 1.725116e-07 3.193295e-07 3.757474e-08 6.999804e-09 4.269611e-08
#>  [6] 5.737161e-08 4.075125e-08 1.365276e-08 1.728020e-08 2.470153e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.122918e+00 ~ 3.022917e+00   [3.308097e-02]
#> 
#> $analytic
#>  [1]  0.4672184  3.1229178  4.5010832  7.4895611  9.0876909 10.6274260
#>  [7] 13.7666684 14.0308833 17.2927976 19.4623148
#> 
#> $finite_difference
#>  [1]  0.4672165  3.0229168  4.5010834  7.4895611  9.0876907 10.6274261
#>  [7] 13.7666686 14.0308833 17.2927979 19.4623150
#> 
#> $relative_error
#>  [1] 4.050472e-06 3.308097e-02 3.279536e-08 8.517935e-09 2.361180e-08
#>  [6] 1.104187e-08 1.329816e-08 2.323182e-10 1.716513e-08 9.288192e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 6.423534e-01 ~ 6.423531e-01   [5.125258e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.619351e+00 ~ 3.519352e+00   [2.841411e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.177199e+00 ~ 5.177199e+00   [2.998155e-10]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.586634e+00 ~ 6.586634e+00   [3.994432e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.265119e+00 ~ 8.265119e+00   [3.370291e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.002130e+01 ~ 1.002130e+01   [2.436429e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.357629e+01 ~ 1.357629e+01   [2.300033e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.509376e+01 ~ 1.509376e+01   [2.128736e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.709520e+01 ~ 1.709520e+01   [5.856450e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.829969e+01 ~ 1.829969e+01   [1.328302e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]    [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.0000  1.00000  1.00000
#> [2,] 0.6423534 3.619351 5.177199 6.586634 8.265119 10.0213 13.57629 15.09376
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.0952 18.29969
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]    [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.0000  1.00000  1.00000
#> [2,] 0.6423531 3.519352 5.177199 6.586634 8.265119 10.0213 13.57629 15.09376
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.0952 18.29969
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 5.125258e-07 0.02841411 2.998155e-10 3.994432e-08 3.370291e-08
#>              [,6]         [,7]         [,8]        [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00 0.000000e+00
#> [2,] 2.436429e-08 2.300033e-08 2.128736e-09 5.85645e-09 1.328302e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>