Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.298519  2.555784  5.453506  6.180207  9.170480 11.932176 13.822965
#>  [8] 15.622431 16.317368 18.455071
#> 
#> $finite_difference
#>  [1]  1.298519  2.555782  5.453506  6.180207  9.170480 11.932176 13.822965
#>  [8] 15.622431 16.317368 18.455071
#> 
#> $relative_error
#>  [1] 1.918561e-07 4.862894e-07 3.244725e-09 1.454201e-08 4.471536e-08
#>  [6] 3.763376e-08 2.709438e-09 8.942167e-11 1.095216e-08 9.689149e-10
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.258777e+00 ~ 2.158777e+00   [4.632235e-02]
#> 
#> $analytic
#>  [1]  1.264819  2.258777  4.733078  7.565412  9.710876 11.286762 13.482448
#>  [8] 15.602810 16.462104 19.564628
#> 
#> $finite_difference
#>  [1]  1.264820  2.158777  4.733078  7.565412  9.710876 11.286762 13.482448
#>  [8] 15.602810 16.462104 19.564628
#> 
#> $relative_error
#>  [1] 4.822949e-07 4.632235e-02 1.121583e-08 2.505140e-08 2.629718e-08
#>  [6] 1.201963e-08 9.956908e-09 6.386767e-09 2.567819e-09 1.070577e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 8.639724e-01 ~ 8.639717e-01   [7.804394e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.686636e+00 ~ 3.586636e+00   [2.788147e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.174304e+00 ~ 4.174304e+00   [1.111824e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.623255e+00 ~ 6.623255e+00   [2.840403e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.525027e+00 ~ 8.525027e+00   [4.406972e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.003683e+01 ~ 1.003683e+01   [5.327720e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.334091e+01 ~ 1.334091e+01   [9.489111e-09]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.439226e+01 ~ 1.439226e+01   [1.908944e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.684125e+01 ~ 1.684125e+01   [1.068520e-08]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.924152e+01 ~ 1.924152e+01   [1.148089e-08]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.8639724 3.686636 4.174304 6.623255 8.525027 10.03683 13.34091 14.39226
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.84125 19.24152
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.8639717 3.586636 4.174304 6.623255 8.525027 10.03683 13.34091 14.39226
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.84125 19.24152
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]        [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00
#> [2,] 7.804394e-07 0.02788147 1.111824e-08 2.840403e-08 4.406972e-08 5.32772e-08
#>              [,7]         [,8]        [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.00000e+00 0.000000e+00
#> [2,] 9.489111e-09 1.908944e-09 1.06852e-08 1.148089e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>