Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.9322612  3.4353392  5.0204086  6.5271011  8.3233770 11.8074972
#>  [7] 13.6996131 15.3502248 16.4117249 19.6749623
#> 
#> $finite_difference
#>  [1]  0.9322624  3.4353409  5.0204086  6.5271015  8.3233772 11.8074977
#>  [7] 13.6996133 15.3502250 16.4117254 19.6749626
#> 
#> $relative_error
#>  [1] 1.276712e-06 4.917763e-07 1.159421e-08 6.456518e-08 3.277328e-08
#>  [6] 3.491557e-08 1.325157e-08 1.101189e-08 2.861960e-08 1.715436e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#> * grad_f[  1 ] = 1.204575e-02 ~ 1.20430e-02   [2.284033e-04]
#> * grad_f[  2 ] = 3.542821e+00 ~ 3.44282e+00   [2.904634e-02]
#> 
#> $analytic
#>  [1]  0.01204575  3.54282089  4.95211997  6.00940807  8.22060502 11.68773942
#>  [7] 12.53303453 14.96563359 16.40215548 19.58579248
#> 
#> $finite_difference
#>  [1]  0.012043  3.442820  4.952119  6.009408  8.220605 11.687739 12.533034
#>  [8] 14.965633 16.402156 19.585793
#> 
#> $relative_error
#>  [1] 2.284033e-04 2.904634e-02 1.578183e-07 1.131335e-08 3.798656e-08
#>  [6] 3.158132e-08 2.699555e-08 1.331738e-08 4.835806e-09 3.067035e-09
#> 
#> $flag_derivative_warning
#>  [1]  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.440136e+00 ~ 1.440136e+00   [1.629661e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.564185e+00 ~ 3.464186e+00   [2.886666e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.945208e+00 ~ 5.945208e+00   [7.936982e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.532461e+00 ~ 7.532461e+00   [2.596465e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 9.516288e+00 ~ 9.516288e+00   [1.734188e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.072965e+01 ~ 1.072965e+01   [3.324399e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.241049e+01 ~ 1.241049e+01   [2.315619e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.509709e+01 ~ 1.509709e+01   [2.615608e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.736443e+01 ~ 1.736443e+01   [7.809694e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.826062e+01 ~ 1.826062e+01   [1.070016e-08]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.440136 3.564185 5.945208 7.532461 9.516288 10.72965 12.41049 15.09709
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.36443 18.26062
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.440136 3.464186 5.945208 7.532461 9.516288 10.72965 12.41049 15.09709
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.36443 18.26062
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.629661e-07 0.02886666 7.936982e-08 2.596465e-08 1.734188e-08
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 3.324399e-08 2.315619e-08 2.615608e-08 7.809694e-09 1.070016e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>