Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.998901  3.218096  4.153400  6.725357  8.969864 10.839924 13.246281
#>  [8] 15.450356 17.102288 18.766902
#> 
#> $finite_difference
#>  [1]  1.998898  3.218096  4.153400  6.725357  8.969864 10.839924 13.246281
#>  [8] 15.450356 17.102288 18.766902
#> 
#> $relative_error
#>  [1] 1.809160e-06 1.481737e-07 1.255696e-08 1.149378e-08 4.121970e-08
#>  [6] 7.531966e-09 1.862164e-08 1.898769e-09 1.845454e-09 7.483571e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 3.713347e+00 ~ 3.613346e+00   [2.767548e-02]
#> 
#> $analytic
#>  [1]  1.221301  3.713347  4.015728  6.491309  8.090006 11.578118 13.708462
#>  [8] 14.910857 17.707518 19.401996
#> 
#> $finite_difference
#>  [1]  1.221302  3.613346  4.015728  6.491309  8.090005 11.578118 13.708462
#>  [8] 14.910857 17.707518 19.401996
#> 
#> $relative_error
#>  [1] 1.194559e-06 2.767548e-02 8.963659e-08 4.354382e-08 1.929414e-08
#>  [6] 1.703948e-08 9.511290e-10 3.747562e-09 1.225750e-08 1.443882e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 3.275239e-01 ~ 3.275261e-01   [6.555704e-06]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.679496e+00 ~ 3.579496e+00   [2.793671e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.679225e+00 ~ 5.679226e+00   [1.715310e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.578534e+00 ~ 6.578535e+00   [1.119818e-07]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 9.835942e+00 ~ 9.835942e+00   [3.713571e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.112244e+01 ~ 1.112244e+01   [4.726765e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.356255e+01 ~ 1.356255e+01   [8.652258e-09]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.573878e+01 ~ 1.573878e+01   [2.559264e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.669696e+01 ~ 1.669696e+01   [2.453611e-08]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.892899e+01 ~ 1.892899e+01   [1.944910e-08]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3275239 3.679496 5.679225 6.578534 9.835942 11.12244 13.56255 15.73878
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.69696 18.92899
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.3275261 3.579496 5.679226 6.578535 9.835942 11.12244 13.56255 15.73878
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.69696 18.92899
#> 
#> $relative_error
#>              [,1]       [,2]        [,3]         [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.00000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 6.555704e-06 0.02793671 1.71531e-07 1.119818e-07 3.713571e-08 4.726765e-08
#>              [,7]         [,8]         [,9]       [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00
#> [2,] 8.652258e-09 2.559264e-08 2.453611e-08 1.94491e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>