Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.916827  2.641294  4.060338  6.024588  9.411040 11.632833 13.462167
#>  [8] 14.847410 17.042245 18.517830
#> 
#> $finite_difference
#>  [1]  1.916828  2.641294  4.060338  6.024589  9.411040 11.632833 13.462167
#>  [8] 14.847410 17.042246 18.517830
#> 
#> $relative_error
#>  [1] 7.581947e-07 1.503842e-07 2.018080e-07 1.041143e-07 1.142006e-08
#>  [6] 9.687238e-09 2.873965e-08 1.210617e-08 2.382344e-08 5.381382e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.945164e+00 ~ 2.845165e+00   [3.514705e-02]
#> 
#> $analytic
#>  [1]  0.9167084  2.9451644  4.6886245  7.8441216  8.5419554 11.9403066
#>  [7] 13.2827548 15.5140028 17.8792598 18.4824690
#> 
#> $finite_difference
#>  [1]  0.9167099  2.8451653  4.6886241  7.8441219  8.5419556 11.9403070
#>  [7] 13.2827552 15.5140028 17.8792602 18.4824692
#> 
#> $relative_error
#>  [1] 1.672238e-06 3.514705e-02 8.756472e-08 4.321725e-08 2.446615e-08
#>  [6] 3.056230e-08 2.941832e-08 1.050544e-09 1.894902e-08 8.334425e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.858687e+00 ~ 1.858688e+00   [7.298003e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.134858e+00 ~ 2.034859e+00   [4.914299e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.339579e+00 ~ 5.339579e+00   [7.383726e-09]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.534325e+00 ~ 7.534326e+00   [2.929571e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.582203e+00 ~ 8.582204e+00   [4.505663e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.095216e+01 ~ 1.095216e+01   [3.826600e-10]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.278861e+01 ~ 1.278861e+01   [9.898911e-09]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.571575e+01 ~ 1.571575e+01   [2.545237e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.760515e+01 ~ 1.760515e+01   [9.651416e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.978963e+01 ~ 1.978963e+01   [1.630669e-08]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.858687 2.134858 5.339579 7.534325 8.582203 10.95216 12.78861 15.71575
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.60515 19.78963
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.858688 2.034859 5.339579 7.534326 8.582204 10.95216 12.78861 15.71575
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.60515 19.78963
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]       [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.0000e+00
#> [2,] 7.298003e-07 0.04914299 7.383726e-09 2.929571e-08 4.505663e-08 3.8266e-10
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 9.898911e-09 2.545237e-08 9.651416e-09 1.630669e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>