Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.2554331  3.0111142  5.9995234  6.2096254  8.0155279 11.9920191
#>  [7] 13.8011776 15.5864290 17.9530392 19.7414300
#> 
#> $finite_difference
#>  [1]  0.2554321  3.0111141  5.9995232  6.2096252  8.0155281 11.9920190
#>  [7] 13.8011774 15.5864291 17.9530394 19.7414303
#> 
#> $relative_error
#>  [1] 3.996089e-06 4.299201e-08 3.725586e-08 2.534670e-08 2.191340e-08
#>  [6] 6.886019e-09 1.279736e-08 5.288065e-09 8.899527e-09 1.427075e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.800712e+00 ~ 2.70071e+00   [3.702797e-02]
#> 
#> $analytic
#>  [1]  1.453806  2.800712  4.719946  7.378167  9.514868 10.177313 12.800556
#>  [8] 14.522648 17.559172 19.577923
#> 
#> $finite_difference
#>  [1]  1.453804  2.700710  4.719945  7.378167  9.514868 10.177312 12.800556
#>  [8] 14.522647 17.559172 19.577923
#> 
#> $relative_error
#>  [1] 1.666549e-06 3.702797e-02 1.194093e-07 2.682320e-08 9.063759e-09
#>  [6] 4.384839e-08 1.382372e-09 1.407631e-08 3.821761e-09 4.847387e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.829437e+00 ~ 1.829437e+00   [3.362447e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.277510e+00 ~ 2.177511e+00   [4.592363e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.809489e+00 ~ 4.809489e+00   [5.660823e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.660096e+00 ~ 7.660096e+00   [3.258372e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 9.743658e+00 ~ 9.743658e+00   [1.468146e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.026038e+01 ~ 1.026038e+01   [2.500681e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.266265e+01 ~ 1.266265e+01   [1.202522e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.551306e+01 ~ 1.551306e+01   [1.008584e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.735866e+01 ~ 1.735866e+01   [1.110890e-08]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.920667e+01 ~ 1.920667e+01   [1.277700e-08]
#> 
#> $analytic
#>          [,1]    [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.01000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.829437 2.27751 4.809489 7.660096 9.743658 10.26038 12.66265 15.51306
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.35866 19.20667
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.829437 2.177511 4.809489 7.660096 9.743658 10.26038 12.66265 15.51306
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.35866 19.20667
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 3.362447e-07 0.04592363 5.660823e-08 3.258372e-08 1.468146e-08
#>              [,6]         [,7]         [,8]        [,9]      [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00 0.0000e+00
#> [2,] 2.500681e-08 1.202522e-08 1.008584e-08 1.11089e-08 1.2777e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>