Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.088406  3.642600  5.666160  6.605610  9.937820 10.039701 12.978362
#>  [8] 15.543328 17.963675 19.759812
#> 
#> $finite_difference
#>  [1]  1.088406  3.642599  5.666161  6.605610  9.937820 10.039701 12.978362
#>  [8] 15.543328 17.963675 19.759813
#> 
#> $relative_error
#>  [1] 7.696806e-07 3.287984e-07 4.881662e-08 7.035383e-08 1.686870e-09
#>  [6] 1.549158e-08 6.048313e-09 1.366127e-08 9.677691e-10 3.737580e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.746405e+00 ~ 2.646404e+00   [3.778733e-02]
#> 
#> $analytic
#>  [1]  1.509068  2.746405  5.568090  7.874577  8.795675 10.552418 13.972097
#>  [8] 15.210143 16.294030 18.221028
#> 
#> $finite_difference
#>  [1]  1.509068  2.646404  5.568090  7.874577  8.795676 10.552418 13.972097
#>  [8] 15.210143 16.294031 18.221028
#> 
#> $relative_error
#>  [1] 5.366128e-07 3.778733e-02 6.919017e-08 4.511983e-08 1.928150e-08
#>  [6] 1.751897e-08 4.380244e-09 6.980261e-09 1.562299e-09 6.148836e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.647914e+00 ~ 1.647911e+00   [1.905980e-06]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.613980e+00 ~ 2.513979e+00   [3.977806e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.700382e+00 ~ 5.700382e+00   [1.486749e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.787046e+00 ~ 7.787045e+00   [1.255789e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.277940e+00 ~ 8.277940e+00   [4.863659e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.036249e+01 ~ 1.036249e+01   [3.319362e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.244270e+01 ~ 1.244270e+01   [1.997932e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.404326e+01 ~ 1.404326e+01   [8.986107e-09]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.766610e+01 ~ 1.766610e+01   [9.562469e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.954565e+01 ~ 1.954565e+01   [6.623153e-10]
#> 
#> $analytic
#>          [,1]    [,2]     [,3]     [,4]    [,5]     [,6]    [,7]     [,8]
#> [1,] 1.000000 1.01000 1.000000 1.000000 1.00000  1.00000  1.0000  1.00000
#> [2,] 1.647914 2.61398 5.700382 7.787046 8.27794 10.36249 12.4427 14.04326
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.6661 19.54565
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]    [,5]     [,6]    [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.00000  1.00000  1.0000  1.00000
#> [2,] 1.647911 2.513979 5.700382 7.787045 8.27794 10.36249 12.4427 14.04326
#>         [,9]    [,10]
#> [1,]  1.0000  1.00000
#> [2,] 17.6661 19.54565
#> 
#> $relative_error
#>             [,1]       [,2]         [,3]         [,4]         [,5]         [,6]
#> [1,] 0.00000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.90598e-06 0.03977806 1.486749e-08 1.255789e-08 4.863659e-08 3.319362e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.997932e-08 8.986107e-09 9.562469e-09 6.623153e-10
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>