Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.6174019  3.2046985  5.1083698  6.0115978  8.0103402 10.6274170
#>  [7] 12.8927409 15.4128493 16.2106437 18.3837658
#> 
#> $finite_difference
#>  [1]  0.6174011  3.2046986  5.1083705  6.0115976  8.0103401 10.6274172
#>  [7] 12.8927411 15.4128494 16.2106438 18.3837658
#> 
#> $relative_error
#>  [1] 1.268612e-06 8.282431e-09 1.297461e-07 1.959750e-08 9.754637e-09
#>  [6] 2.621716e-08 1.229563e-08 7.764622e-09 3.619432e-09 1.185445e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.921312e+00 ~ 2.821312e+00   [3.544458e-02]
#> 
#> $analytic
#>  [1]  0.1120646  2.9213122  5.0636445  7.9450703  8.2928169 11.0551340
#>  [7] 13.2789488 15.5907680 17.7003297 19.4824882
#> 
#> $finite_difference
#>  [1]  0.1120644  2.8213120  5.0636444  7.9450703  8.2928169 11.0551338
#>  [7] 13.2789486 15.5907679 17.7003297 19.4824883
#> 
#> $relative_error
#>  [1] 2.040253e-06 3.544458e-02 2.758851e-08 5.509349e-09 9.882815e-10
#>  [6] 1.625899e-08 1.120661e-08 6.511171e-09 3.250507e-09 3.604351e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 7.217224e-01 ~ 7.217216e-01   [1.080725e-06]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 3.810935e+00 ~ 3.710934e+00   [2.694773e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 5.319642e+00 ~ 5.319641e+00   [1.510000e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 6.245353e+00 ~ 6.245353e+00   [3.504379e-09]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.476053e+00 ~ 8.476052e+00   [6.238812e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.188876e+01 ~ 1.188876e+01   [2.223448e-08]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.382147e+01 ~ 1.382147e+01   [1.407809e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.550136e+01 ~ 1.550136e+01   [3.905205e-10]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.784711e+01 ~ 1.784711e+01   [1.128901e-08]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.861124e+01 ~ 1.861124e+01   [1.401144e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.7217224 3.810935 5.319642 6.245353 8.476053 11.88876 13.82147 15.50136
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.84711 18.61124
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.7217216 3.710934 5.319641 6.245353 8.476052 11.88876 13.82147 15.50136
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.84711 18.61124
#> 
#> $relative_error
#>              [,1]       [,2]     [,3]         [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.01000000 0.00e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.080725e-06 0.02694773 1.51e-07 3.504379e-09 6.238812e-08 2.223448e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.407809e-08 3.905205e-10 1.128901e-08 1.401144e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>