Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.8061264  3.3598735  4.7536379  7.5314704  9.9029937 10.4878141
#>  [7] 13.6334937 15.3500128 16.2601801 19.2520414
#> 
#> $finite_difference
#>  [1]  0.8061256  3.3598747  4.7536379  7.5314703  9.9029938 10.4878146
#>  [7] 13.6334937 15.3500133 16.2601802 19.2520416
#> 
#> $relative_error
#>  [1] 9.288668e-07 3.785025e-07 4.146928e-09 8.594194e-09 8.454605e-09
#>  [6] 4.272781e-08 8.831684e-10 2.924411e-08 1.011881e-09 1.055548e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 2.371624e+00 ~ 2.271624e+00   [4.402145e-02]
#> 
#> $analytic
#>  [1]  0.884096  2.371624  5.543150  6.165714  8.185518 11.460453 12.125435
#>  [8] 15.695074 17.324859 18.708667
#> 
#> $finite_difference
#>  [1]  0.8840981  2.2716236  5.5431506  6.1657152  8.1855186 11.4604530
#>  [7] 12.1254354 15.6950746 17.3248596 18.7086678
#> 
#> $relative_error
#>  [1] 2.295919e-06 4.402145e-02 7.260971e-08 1.255969e-07 5.126785e-08
#>  [6] 2.868620e-08 2.059534e-08 7.981032e-09 2.523864e-08 1.904095e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 1.699261e+00 ~ 1.699265e+00   [2.129267e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 3.342835e+00 ~ 3.242836e+00   [3.083692e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 4.138814e+00 ~ 4.138814e+00   [5.273006e-08]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 7.051631e+00 ~ 7.051631e+00   [1.994286e-08]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 9.639038e+00 ~ 9.639038e+00   [5.792361e-08]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.181546e+01 ~ 1.181546e+01   [4.207800e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.227812e+01 ~ 1.227812e+01   [1.170833e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.413821e+01 ~ 1.413821e+01   [2.904987e-08]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.678774e+01 ~ 1.678774e+01   [2.539275e-08]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.951466e+01 ~ 1.951467e+01   [1.541016e-08]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.699261 3.342835 4.138814 7.051631 9.639038 11.81546 12.27812 14.13821
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.78774 19.51466
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.699265 3.242836 4.138814 7.051631 9.639038 11.81546 12.27812 14.13821
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.78774 19.51467
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]       [,6]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.0000e+00
#> [2,] 2.129267e-06 0.03083692 5.273006e-08 1.994286e-08 5.792361e-08 4.2078e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.170833e-08 2.904987e-08 2.539275e-08 1.541016e-08
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>