Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.964843  2.799016  4.933021  7.093819  8.702093 11.459198 12.593098
#>  [8] 14.787624 16.370853 19.145361
#> 
#> $finite_difference
#>  [1]  1.964844  2.799015  4.933020  7.093819  8.702093 11.459198 12.593098
#>  [8] 14.787624 16.370853 19.145362
#> 
#> $relative_error
#>  [1] 2.301237e-07 4.200745e-07 1.005011e-07 6.072000e-08 2.235707e-08
#>  [6] 2.560097e-09 1.951355e-08 6.927785e-10 3.685148e-09 1.445724e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 2.3545e+00 ~ 2.254499e+00   [4.435592e-02]
#> 
#> $analytic
#>  [1]  0.6257678  2.3544998  5.9708987  6.8877074  8.5040098 11.0610194
#>  [7] 13.4565712 14.5015385 16.1572227 18.7856113
#> 
#> $finite_difference
#>  [1]  0.6257668  2.2544994  5.9708989  6.8877068  8.5040092 11.0610193
#>  [7] 13.4565713 14.5015383 16.1572227 18.7856113
#> 
#> $relative_error
#>  [1] 1.645305e-06 4.435592e-02 4.466140e-08 9.661125e-08 6.006936e-08
#>  [6] 1.639066e-08 8.715454e-09 1.213802e-08 5.389457e-09 4.451959e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.089909e+00 ~ 1.089909e+00   [3.204360e-08]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.499787e+00 ~ 2.399786e+00   [4.167092e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.300640e+00 ~ 4.300640e+00   [7.655198e-08]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.408096e+00 ~ 7.408096e+00   [1.012021e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 9.203699e+00 ~ 9.203698e+00   [6.231283e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.104181e+01 ~ 1.104181e+01   [3.106710e-09]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.243981e+01 ~ 1.243981e+01   [3.208555e-10]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.433618e+01 ~ 1.433618e+01   [1.201818e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.708905e+01 ~ 1.708905e+01   [6.984854e-09]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.824127e+01 ~ 1.824127e+01   [8.112767e-09]
#> 
#> $analytic
#>          [,1]     [,2]    [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.010000 1.00000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.089909 2.499787 4.30064 7.408096 9.203699 11.04181 12.43981 14.33618
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.08905 18.24127
#> 
#> $finite_difference
#>          [,1]     [,2]    [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.000000 1.000000 1.00000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 1.089909 2.399786 4.30064 7.408096 9.203698 11.04181 12.43981 14.33618
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.08905 18.24127
#> 
#> $relative_error
#>             [,1]       [,2]         [,3]         [,4]         [,5]        [,6]
#> [1,] 0.00000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00
#> [2,] 3.20436e-08 0.04167092 7.655198e-08 1.012021e-08 6.231283e-08 3.10671e-09
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 3.208555e-10 1.201818e-08 6.984854e-09 8.112767e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>