Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function(x, a) sum((x - a) ^ 2)

f_grad <- function(x, a)  2 * (x - a)

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'none', a = runif(10))
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  1.933511  2.813773  5.948133  7.376948  8.160420 11.771038 12.541676
#>  [8] 14.377246 16.019119 18.136797
#> 
#> $finite_difference
#>  [1]  1.933514  2.813774  5.948134  7.376948  8.160420 11.771038 12.541677
#>  [8] 14.377246 16.019120 18.136797
#> 
#> $relative_error
#>  [1] 1.191179e-06 3.652441e-07 1.352277e-07 1.868466e-08 1.406042e-08
#>  [6] 7.160345e-09 1.847438e-08 2.688268e-09 1.692467e-08 1.804949e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function(x, a)  2 * (x - a) + c(0, 0.1, rep(0, 8))

check.derivatives(.x = 1:10, func = f, func_grad = f_grad,
          check_derivatives_print = 'errors', a = runif(10))
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[ 2] = 2.565981e+00 ~ 2.465981e+00   [4.055208e-02]
#> 
#> $analytic
#>  [1]  1.397902  2.565981  5.326895  7.010927  8.395821 10.611325 13.022024
#>  [8] 14.598651 17.112536 18.896966
#> 
#> $finite_difference
#>  [1]  1.397900  2.465981  5.326894  7.010927  8.395821 10.611324 13.022024
#>  [8] 14.598650 17.112537 18.896966
#> 
#> $relative_error
#>  [1] 1.616274e-06 4.055208e-02 9.531371e-08 3.208055e-08 4.364975e-08
#>  [6] 2.913860e-08 1.551452e-08 3.317346e-09 5.511882e-09 3.302044e-10
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function(x, a) c(sum(x - a), sum((x - a) ^ 2))

g_grad <- function(x, a) {
   rbind(rep(1, length(x)) + c(0, 0.01, rep(0, 8)),
   2 * (x - a) + c(0, 0.1, rep(0, 8)))
}

check.derivatives(.x = 1:10, func = g, func_grad = g_grad,
          check_derivatives_print = 'all', a = runif(10))
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[1,  1] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  1] = 8.055132e-01 ~ 8.055115e-01   [2.147041e-06]
#> * grad_f[1,  2] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[2,  2] = 2.738956e+00 ~ 2.638956e+00   [3.789390e-02]
#>   grad_f[1,  3] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  3] = 5.723377e+00 ~ 5.723376e+00   [1.704248e-07]
#>   grad_f[1,  4] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  4] = 7.002892e+00 ~ 7.002892e+00   [3.131941e-08]
#>   grad_f[1,  5] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  5] = 9.574745e+00 ~ 9.574745e+00   [6.808806e-10]
#>   grad_f[1,  6] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  6] = 1.001784e+01 ~ 1.001784e+01   [4.073476e-08]
#>   grad_f[1,  7] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  7] = 1.266159e+01 ~ 1.266159e+01   [1.186862e-08]
#>   grad_f[1,  8] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  8] = 1.503664e+01 ~ 1.503664e+01   [1.192285e-08]
#>   grad_f[1,  9] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2,  9] = 1.781195e+01 ~ 1.781195e+01   [7.476952e-09]
#>   grad_f[1, 10] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[2, 10] = 1.911726e+01 ~ 1.911726e+01   [6.464656e-09]
#> 
#> $analytic
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.8055132 2.738956 5.723377 7.002892 9.574745 10.01784 12.66159 15.03664
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.81195 19.11726
#> 
#> $finite_difference
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 1.0000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.00000
#> [2,] 0.8055115 2.638956 5.723376 7.002892 9.574745 10.01784 12.66159 15.03664
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 17.81195 19.11726
#> 
#> $relative_error
#>              [,1]      [,2]         [,3]         [,4]         [,5]         [,6]
#> [1,] 0.000000e+00 0.0100000 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 2.147041e-06 0.0378939 1.704248e-07 3.131941e-08 6.808806e-10 4.073476e-08
#>              [,7]         [,8]         [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 1.186862e-08 1.192285e-08 7.476952e-09 6.464656e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>