Check analytic gradients of a function using finite difference approximations

This function compares the analytic gradients of a function with a finite difference approximation and prints the results of these checks.

check.derivatives(
  .x,
  func,
  func_grad,
  check_derivatives_tol = 1e-04,
  check_derivatives_print = "all",
  func_grad_name = "grad_f",
  ...
)

Arguments

.x

point at which the comparison is done.

func

function to be evaluated.

func_grad

function calculating the analytic gradients.

check_derivatives_tol

option determining when differences between the analytic gradient and its finite difference approximation are flagged as an error.

check_derivatives_print

option related to the amount of output. 'all' means that all comparisons are shown, 'errors' only shows comparisons that are flagged as an error, and 'none' shows the number of errors only.

func_grad_name

option to change the name of the gradient function that shows up in the output.

...

further arguments passed to the functions func and func_grad.

Value

The return value contains a list with the analytic gradient, its finite difference approximation, the relative errors, and vector comparing the relative errors to the tolerance.

See also

Author

Jelmer Ypma

Examples

library('nloptr')

# example with correct gradient
f <- function( x, a ) {
  return( sum( ( x - a )^2 ) )
}

f_grad <- function( x, a ) {
  return( 2*( x - a ) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='none', a=runif(10) )
#> Derivative checker results: 0 error(s) detected.
#> $analytic
#>  [1]  0.5579855  3.2224453  4.4683987  6.8559133  9.7337995 10.8466275
#>  [7] 13.4346778 15.6207322 17.5013836 18.1585482
#> 
#> $finite_difference
#>  [1]  0.5579834  3.2224445  4.4683978  6.8559132  9.7337990 10.8466276
#>  [7] 13.4346777 15.6207323 17.5013835 18.1585484
#> 
#> $relative_error
#>  [1] 3.778812e-06 2.395903e-07 2.000526e-07 2.581002e-08 4.975321e-08
#>  [6] 2.046398e-09 6.471731e-09 9.002760e-09 8.638443e-09 6.872643e-09
#> 
#> $flag_derivative_warning
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient
f_grad <- function( x, a ) {
  return( 2*( x - a ) + c(0,.1,rep(0,8)) )
}

check.derivatives( .x=1:10, func=f, func_grad=f_grad,
    check_derivatives_print='errors', a=runif(10) )
#> Derivative checker results: 1 error(s) detected.
#> 
#> * grad_f[  2 ] = 4.041012e+00 ~ 3.941011e+00   [2.537426e-02]
#> 
#> $analytic
#>  [1]  0.5134712  4.0410117  5.0112966  6.2401244  8.7141524 10.1555377
#>  [7] 12.6744310 14.9631545 16.8659492 18.2136717
#> 
#> $finite_difference
#>  [1]  0.5134697  3.9410114  5.0112966  6.2401247  8.7141525 10.1555379
#>  [7] 12.6744308 14.9631548 16.8659494 18.2136719
#> 
#> $relative_error
#>  [1] 2.953746e-06 2.537426e-02 8.672750e-09 5.544542e-08 1.900232e-08
#>  [6] 1.860101e-08 1.583503e-08 1.745862e-08 1.399497e-08 1.177597e-08
#> 
#> $flag_derivative_warning
#>  [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> 

# example with incorrect gradient of vector-valued function
g <- function( x, a ) {
  return( c( sum(x-a), sum( (x-a)^2 ) ) )
}

g_grad <- function( x, a ) {
  return( rbind( rep(1,length(x)) + c(0,.01,rep(0,8)), 2*(x-a) + c(0,.1,rep(0,8)) ) )
}

check.derivatives( .x=1:10, func=g, func_grad=g_grad,
    check_derivatives_print='all', a=runif(10) )
#> Derivative checker results: 2 error(s) detected.
#> 
#>   grad_f[ 1,  1 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  1 ] = 1.147178e+00 ~ 1.147179e+00   [9.348308e-07]
#> * grad_f[ 1,  2 ] = 1.010000e+00 ~ 1.000000e+00   [1.000000e-02]
#> * grad_f[ 2,  2 ] = 2.589192e+00 ~ 2.489193e+00   [4.017332e-02]
#>   grad_f[ 1,  3 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  3 ] = 4.361251e+00 ~ 4.361252e+00   [1.142465e-07]
#>   grad_f[ 1,  4 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  4 ] = 7.922545e+00 ~ 7.922545e+00   [6.418419e-08]
#>   grad_f[ 1,  5 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  5 ] = 8.418058e+00 ~ 8.418058e+00   [2.742614e-08]
#>   grad_f[ 1,  6 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  6 ] = 1.135868e+01 ~ 1.135868e+01   [3.074705e-09]
#>   grad_f[ 1,  7 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  7 ] = 1.378384e+01 ~ 1.378384e+01   [1.713287e-08]
#>   grad_f[ 1,  8 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  8 ] = 1.507400e+01 ~ 1.507400e+01   [1.235668e-08]
#>   grad_f[ 1,  9 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2,  9 ] = 1.627618e+01 ~ 1.627618e+01   [6.675670e-11]
#>   grad_f[ 1, 10 ] = 1.000000e+00 ~ 1.000000e+00   [0.000000e+00]
#>   grad_f[ 2, 10 ] = 1.957436e+01 ~ 1.957436e+01   [4.263052e-09]
#> 
#> $analytic
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]   [,8]
#> [1,] 1.000000 1.010000 1.000000 1.000000 1.000000  1.00000  1.00000  1.000
#> [2,] 1.147178 2.589192 4.361251 7.922545 8.418058 11.35868 13.78384 15.074
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.27618 19.57436
#> 
#> $finite_difference
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]   [,8]
#> [1,] 1.000000 1.000000 1.000000 1.000000 1.000000  1.00000  1.00000  1.000
#> [2,] 1.147179 2.489193 4.361252 7.922545 8.418058 11.35868 13.78384 15.074
#>          [,9]    [,10]
#> [1,]  1.00000  1.00000
#> [2,] 16.27618 19.57436
#> 
#> $relative_error
#>              [,1]       [,2]         [,3]         [,4]         [,5]
#> [1,] 0.000000e+00 0.01000000 0.000000e+00 0.000000e+00 0.000000e+00
#> [2,] 9.348308e-07 0.04017332 1.142465e-07 6.418419e-08 2.742614e-08
#>              [,6]         [,7]         [,8]        [,9]        [,10]
#> [1,] 0.000000e+00 0.000000e+00 0.000000e+00 0.00000e+00 0.000000e+00
#> [2,] 3.074705e-09 1.713287e-08 1.235668e-08 6.67567e-11 4.263052e-09
#> 
#> $flag_derivative_warning
#>       [,1] [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10]
#> [1,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>