isres {nloptr}	R Documentation

Improved Stochastic Ranking Evolution Strategy

Description

The Improved Stochastic Ranking Evolution Strategy (ISRES) algorithm for nonlinearly constrained global optimization (or at least semi-global: although it has heuristics to escape local optima.

Usage

isres(
  x0,
  fn,
  lower,
  upper,
  hin = NULL,
  heq = NULL,
  maxeval = 10000,
  pop.size = 20 * (length(x0) + 1),
  xtol_rel = 1e-06,
  nl.info = FALSE,
  ...
)

Arguments

x0
initial point for searching the optimum.

fn
objective function that is to be minimized.

lower, upper
lower and upper bound constraints.

hin
function defining the inequality constraints, that is hin>=0 for all components.

heq
function defining the equality constraints, that is heq==0 for all components.

maxeval
maximum number of function evaluations.

pop.size
population size.

xtol_rel
stopping criterion for relative change reached.

nl.info
logical; shall the original NLopt info been shown.

...
additional arguments passed to the function.

Details

The evolution strategy is based on a combination of a mutation rule (with a log-normal step-size update and exponential smoothing) and differential variation (a Nelder-Mead-like update rule). The fitness ranking is simply via the objective function for problems without nonlinear constraints, but when nonlinear constraints are included the stochastic ranking proposed by Runarsson and Yao is employed.

This method supports arbitrary nonlinear inequality and equality constraints in addition to the bound constraints.

Value

List with components:

par
the optimal solution found so far.

value
the function value corresponding to par.

iter
number of (outer) iterations, see maxeval.

convergence
integer code indicating successful completion (> 0) or a possible error number (< 0).

message
character string produced by NLopt and giving additional information.

Note

The initial population size for CRS defaults to 20x(n+1) in n dimensions, but this can be changed; the initial population must be at least n+1.

Author(s)

Hans W. Borchers

References

Thomas Philip Runarsson and Xin Yao, “Search biases in constrained evolutionary optimization,” IEEE Trans. on Systems, Man, and Cybernetics Part C: Applications and Reviews, vol. 35 (no. 2), pp. 233-243 (2005).

Examples

### Rosenbrock Banana objective function
fn <- function(x)
    return( 100 * (x[2] - x[1] * x[1])^2 + (1 - x[1])^2 )

x0 <- c( -1.2, 1 )
lb <- c( -3, -3 )
ub <- c(  3,  3 )

isres(x0 = x0, fn = fn, lower = lb, upper = ub)

[Package nloptr version 2.0.3 Index]