newuoa {nloptr}	R Documentation

New Unconstrained Optimization with quadratic Approximation

Description

NEWUOA solves quadratic subproblems in a spherical trust regionvia a truncated conjugate-gradient algorithm. For bound-constrained problems, BOBYQA shold be used instead, as Powell developed it as an enhancement thereof for bound constraints.

Usage

newuoa(x0, fn, nl.info = FALSE, control = list(), ...)

Arguments

x0
starting point for searching the optimum.

fn
objective function that is to be minimized.

nl.info
logical; shall the original NLopt info been shown.

control
list of options, see nl.opts for help.

...
additional arguments passed to the function.

Details

This is an algorithm derived from the NEWUOA Fortran subroutine of Powell, converted to C and modified for the NLOPT stopping criteria.

Value

List with components:

par
the optimal solution found so far.

value
the function value corresponding to par.

iter
number of (outer) iterations, see maxeval.

convergence
integer code indicating successful completion (> 0) or a possible error number (< 0).

message
character string produced by NLopt and giving additional information.

Note

NEWUOA may be largely superseded by BOBYQA.

Author(s)

Hans W. Borchers

References

M. J. D. Powell. “The BOBYQA algorithm for bound constrained optimization without derivatives,” Department of Applied Mathematics and Theoretical Physics, Cambridge England, technical reportNA2009/06 (2009).

See Also

bobyqa, cobyla

Examples

fr <- function(x) {   ## Rosenbrock Banana function
    100 * (x[2] - x[1]^2)^2 + (1 - x[1])^2
}
(S <- newuoa(c(1, 2), fr))

[Package nloptr version 2.0.3 Index]