nl.opts {nloptr}	R Documentation

Setting NL Options

Description

Sets and changes the NLOPT options.

Usage

nl.opts(optlist = NULL)

Arguments

optlist
list of options, see below.

Details

The following options can be set (here with default values):

stopval = -Inf, # stop minimization at this value
xtol_rel = 1e-6, # stop on small optimization step
maxeval = 1000, # stop on this many function evaluations
ftol_rel = 0.0, # stop on change times function value
ftol_abs = 0.0, # stop on small change of function value
check_derivatives = FALSE

Value

returns a list with default and changed options.

Note

There are more options that can be set for solvers in NLOPT. These cannot be set through their wrapper functions. To see the full list of options and algorithms, type nloptr.print.options().

Author(s)

Hans W. Borchers

Examples

nl.opts(list(xtol_rel = 1e-8, maxeval = 2000))

[Package nloptr version 2.0.3 Index]