nloptr.print.options {nloptr}	R Documentation

Print description of nloptr options

Description

This function prints a list of all the options that can be set when solving a minimization problem using nloptr.

Usage

nloptr.print.options(opts.show = NULL, opts.user = NULL)

Arguments

opts.show
list or vector with names of options. A description will be shown for the options in this list. By default, a description of all options is shown.

opts.user
object containing user supplied options. This argument is optional. It is used when nloptr.print.options is called from nloptr. In that case options are listed if print_options_doc is set to TRUE when passing a minimization problem to nloptr.

Author(s)

Jelmer Ypma

See Also

nloptr

Examples

library('nloptr')
nloptr.print.options()

nloptr.print.options( opts.show = c("algorithm", "check_derivatives") )

opts <- list("algorithm"="NLOPT_LD_LBFGS",
             "xtol_rel"=1.0e-8)
nloptr.print.options( opts.user = opts )

[Package nloptr version 2.0.3 Index]