print.nloptr {nloptr}	R Documentation

Print results after running nloptr

Description

This function prints the nloptr object that holds the results from a minimization using nloptr.

Usage

## S3 method for class 'nloptr'
print(x, show.controls = TRUE, ...)

Arguments

x
object containing result from minimization.

show.controls
Logical or vector with indices. Should we show the value of the control variables in the solution? If show.controls is a vector with indices, it is used to select which control variables should be shown. This can be useful if the model contains a set of parameters of interest and a set of nuisance parameters that are not of immediate interest.

...
further arguments passed to or from other methods.

Author(s)

Jelmer Ypma

See Also

nloptr

[Package nloptr version 2.0.3 Index]